Added a basic UI layout
Added the "stage" with JavaFX in GameApplication.java, as well as a layout (BoardView.fxml) for the UI using scene builder. Also added GameController.java for future use of connecting the UI controls to the actual game logic we will implement later.
Also note: the project should be run with Maven in order to display the window + UI